==> 1_News_Data_Collection.ipynb <==
!pip install intrinio_sdk

import pandas as pd
import nltk
import numpy as np
nltk.download('wordnet')


# API call to Intrinio for company news data (https://intrinio.com/)
#Since Github Does not accept files larger than 25 MB, I put an error to reduce its size.
from __future__ import print_function
import time
import intrinio_sdk
from intrinio_sdk.rest import ApiException
from pprint import pprint

intrinio_sdk.ApiClient().configuration.api_key['api_key'] = 'OjgxN2FiYThhMzc3MjY4NzA1NDY4MThhNDQxMDNmZGQ3'

company_api = intrinio_sdk.CompanyApi()

identifier = 'AAPL' # str | A Company identifier (Ticker, CIK, LEI, Intrinio ID)
page_size = 100 # int | The number of results to return (optional) (default to 100)
next_page = '' # str | Gets the next page of data from a previous API call (optional)

api_response_list = []
condition = True
while condition:

    try:
        api_response = company_api.get_company_news(identifier, page_size=page_size, next_page=next_page)
        api_response_list.append(api_response)
        next_page = api_response.next_page
        time.sleep(5)
    except ApiException as e:
        print("Exception when calling CompanyApi->get_company_news: %s\r\n" % e)
        condition = False

#Extract URL, Title, Date and Summary of each piece of News
url_list = []
title_list = []
date_list = []
summary_list = []

cnt = 0
last = ''

for page_100 in api_response_list:
    for article in page_100.news:
        
        if last == article.title:
            cnt+= 1
            pass 
        last = article.title
        
        
        url_list.append(article.url)
        title_list.append(article.title)
        date_list.append(article.publication_date)
        summary_list.append(article.summary)

#Assigning news to dates within stock trading hours
#If news was released after 4pm (when the stock market closes), the news affects next day's stock price
from datetime import timedelta
new_date = []
for date_item in date_list:
    if date_item.hour < 15:
        new_date.append(date_item)
    elif date_item.hour > 16:
        new_date.append(date_item - timedelta(days=1))
    else:
        new_date.append(np.nan)

#Dataframe with each New's URL, Title, Date and Summary
new_df = pd.DataFrame ({'url':url_list,
                       'title':title_list,
                       'date':new_date,
                       'summary':summary_list})

new_df.head()

#Extracting the News Headlines and Date
Int_titles = new_df['date','title']
Int_titles = Int_titles.to_frame()

Int_titles.to_csv('/content/drive/MyDrive/Colab Notebooks/BERT + STOCK /Intrinio_Titles_Data.csv')
