==> 2_Market_Data_Collection.ipynb <==
!pip install intrinio_sdk

import pandas as pd
import nltk
import numpy as np
nltk.download('wordnet')

from __future__ import print_function
import time
import intrinio_sdk
from intrinio_sdk.rest import ApiException
from pprint import pprint

intrinio_sdk.ApiClient().configuration.api_key['api_key'] = 'OjhhMjAzZjE5MmFkZmRkMGQ1YTgzZDU4ODFmMmJiYTJm'

security_api = intrinio_sdk.SecurityApi()

identifier = 'AAPL' # str | A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)
start_date = '2015-01-01' # date | Return prices on or after the date (optional)
end_date = '2020-12-31' # date | Return prices on or before the date (optional)
frequency = 'daily' # str | Return stock prices in the given frequency (optional) (default to daily)
page_size = 100 # int | The number of results to return (optional) (default to 100)
next_page = '' # str | Gets the next page of data from a previous API call (optional)

api_response_list_1 = []
condition = True

while condition:
    try:
        api_response = security_api.get_security_stock_prices(identifier, start_date=start_date, end_date=end_date, frequency=frequency, page_size=page_size, next_page=next_page)
        #pprint(api_response)
        api_response_list_1.append(api_response)
        next_page = api_response.next_page
    except ApiException as e:
        print("Exception when calling SecurityApi->get_security_stock_prices: %s\r\n" % e)
        condition = False

#Extract close, high, low and open of each piece of stock data
close_list = []
high_list = []
low_list = []
open_list = []
date_list = []

cnt = 0
last = ''

for page_100 in api_response_list_1:
    for price in page_100.stock_prices:
        if last == price.open:
            cnt+= 1
            pass 
        last = price.open
        

        close_list.append(price.close)
        high_list.append(price.high)
        low_list.append(price.low)
        open_list.append(price.open)
        date_list.append(price.date)

new_df = pd.DataFrame ({'close':close_list,
                       'high':high_list,
                       'low':low_list,
                       'open':open_list,
                       'date':date_list
                       })

new_df.head()

df_1.to_csv('/content/drive/MyDrive/Colab Notebooks/BERT + STOCK /Market_Data_apple.csv')
