==> 3)_Merging_Market_and_News_Data_Collected_from_Intrinio.ipynb <==
import pandas as pd
import nltk
import numpy as np

stock_data = pd.read_csv("/content/drive/MyDrive/BERT + STOCK /Market_Data_apple.csv")

stock_data.head()

stock_data.shape

news_data = pd.read_csv("/content/drive/MyDrive/BERT + STOCK /Intrinio_Titles_Data.csv")

news_data.head()

news_data.shape

news_data['date']=pd.to_datetime(news_data['date']).dt.strftime('%Y-%m-%d')

stock_data['date']=pd.to_datetime(stock_data['date']).dt.strftime('%Y-%m-%d')

news_data.head()

stock_data.head()

news_data.isna().sum()

news_data.dropna(inplace=True)

news_data.columns

news_data = news_data[['date','title']]

stock_data.columns

stock_data = stock_data[['close', 'high', 'low', 'open', 'date', 'difference']]

stock_data.head()

news_data.head()

news_data.shape

stock_data.shape

#Merging News and Stock data to one table
result = pd.merge(news_data, stock_data, on='date')
result.head()

result.head()

result.shape

result.drop_duplicates(inplace=True)

result.shape

result.head(10)

result.reset_index(inplace=True)

result.head(10)

result.columns

result = result[['date', 'title', 'close', 'high', 'low', 'open','difference']]

#result.to_csv('/content/drive/MyDrive/Merged_Apple_News_difference.csv')
