==> 5)_Labelling_using_MACD.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('/content/drive/MyDrive/BERT + STOCK /Merged_App_News_MACD.csv')

data.head()

stock_data = pd.read_csv("/content/drive/MyDrive/BERT + STOCK /Market_Data_apple.csv")

stock_data.head()

data_sample = data.iloc[0:5585]

data_sample.head()

data.columns

macd = data_sample[['macd','macds']]

macd_signal = data[['macds']]

macd.shape

macd_list = macd.macd.to_list()

macd_signal_list = macd_signal.macds.to_list()

macd['Difference'] = macd['macd'] - macd['macds']

macd.head()

macd_list_difference = macd.Difference.to_list()

len(macd_list_difference)

label_alt_df = pd.DataFrame(label_alt)
label_alt_df = label_alt_df.rename(columns={0:"Label_Signal"})
label_alt_df.head()

Bullish_data = label_alt_df.loc[label_alt_df['Label_Signal'] == "Bullish"] 
Bearish_data = label_alt_df.loc[label_alt_df['Label_Signal'] == "Bearish"] 
Hold_data = label_alt_df.loc[label_alt_df['Label_Signal'] == "Hold"] 

label_signal_df = pd.DataFrame(label_signal)
label_signal_df = label_signal_df.rename(columns={0:"Label_Signal"})
label_signal_df.head()

Bullish_data = label_signal_df.loc[label_signal_df['Label_Signal'] == "Bullish"] 
Bullish_data.shape

Bearish_data = label_signal_df.loc[label_signal_df['Label_Signal'] == "Bearish"] 
Bearish_data.shape

Hold_data = label_signal_df.loc[label_signal_df['Label_Signal'] == "Hold"] 
Hold_data.shape

len(macd_list_difference)

#for macd label
label = []
index = 0
for i in range(len(macd_list_difference)-1):
  if i==5585:
    break
  current_index = i
  next_index = current_index + 1
  if (macd_list_difference[i] < 0.00000 and macd_list_difference[next_index]>0.000000):
    label.append('Bearish')
  elif (macd_list_difference[i] > 0.000000 and macd_list_difference[next_index]<0.000000):
    label.append('Bullish')
  else:
    label.append('Hold')  


label_df = pd.DataFrame(label)
label_df = label_df.rename(columns={0:"Label"})
label_df.head()

label_data = pd.merge(data, label_df, left_index=True, right_index=True)

label_data.shape

label_data.head(10)

Bullish_data = label_data.loc[label_data['Label'] == "Bullish"] 
Bullish_data.shape

Bearish_data = label_data.loc[label_data['Label'] == "Bearish"] 
Bearish_data.shape

Hold_data = label_data.loc[label_data['Label'] == "Hold"]
Hold_data.shape

label_data.head(20)

#label_data.to_csv()
