==> 6)_Generating_Sentence_Vectors_using_BERT.ipynb <==
# Install the library using pip
!pip install sentence-transformers

from sentence_transformers import SentenceTransformer

# Load the BERT model. Various models trained on Natural Language Inference (NLI) https://github.com/UKPLab/sentence-transformers/blob/master/docs/pretrained-models/nli-models.md and 
# Semantic Textual Similarity are available https://github.com/UKPLab/sentence-transformers/blob/master/docs/pretrained-models/sts-models.md

model = SentenceTransformer('bert-base-nli-mean-tokens')

import pandas as pd

data = pd.read_csv('/content/drive/MyDrive/BERT + STOCK /Label_data_march_02 (1).csv')

data.head()

data.shape

sentences = data.title.to_list()

sentence_embeddings = model.encode(sentences)

print('Sample BERT embedding vector - length', len(sentence_embeddings[0]))

print('Sample BERT embedding vector - note includes negative values', sentence_embeddings[0])

title_embeddings = pd.DataFrame(sentence_embeddings)

title_embeddings.shape

title_embeddings.to_csv('/content/drive/MyDrive/BERT + STOCK /title_embeddings.csv', index=False)
