==> 8)_Model_Comparison_using_T_Test.ipynb <==
from scipy import stats

LSTM= [0.8376480501849164, 0.8376480501849276]
MLP = [0.8850709236458967, 0.8750709236458967]
XGBoost = [0.8926548382566359, 0.8726548382566359]

stats.ttest_rel(LSTM,MLP)

stats.ttest_rel(LSTM,XGBoost)

stats.ttest_rel(MLP,XGBoost)
